Moved SessionSimulation stub to stub classes package
